Move start all/stop all commands to TestShop
