+++ playground/Stress/Stress.AppHost/Program.cs
@@ -95,25 +95,7 @@ builder.AddProject<Projects.Stress_TelemetryService>("stress-telemetryservice")
            {
                return Task.FromResult(CommandResults.Success());
            },
-           commandOptions: new() { IconName = "CloudDatabase" })
-       .WithCommand(
-           name: "resource-stop-all",
-           displayName: "Stop all resources",
-           executeCommand: async (c) =>
-           {
-               await ExecuteCommandForAllResourcesAsync(c.ServiceProvider, "resource-stop", c.CancellationToken);
-               return CommandResults.Success();
-           },
-           commandOptions: new() { IconName = "Stop", IconVariant = IconVariant.Filled })
-       .WithCommand(
-           name: "resource-start-all",
-           displayName: "Start all resources",
-           executeCommand: async (c) =>
-           {
-               await ExecuteCommandForAllResourcesAsync(c.ServiceProvider, "resource-start", c.CancellationToken);
-               return CommandResults.Success();
-           },
-           commandOptions: new() { IconName = "Play", IconVariant = IconVariant.Filled });
+           commandOptions: new() { IconName = "CloudDatabase" });
 
 #if !SKIP_DASHBOARD_REFERENCE
 // This project is only added in playground projects to support development/debugging
@@ -140,21 +122,3 @@ for (var i = 0; i < 3; i++)
 }
 
 builder.Build().Run();
-
-static async Task ExecuteCommandForAllResourcesAsync(IServiceProvider serviceProvider, string commandName, CancellationToken cancellationToken)
-{
-    var commandService = serviceProvider.GetRequiredService<ResourceCommandService>();
-    var model = serviceProvider.GetRequiredService<DistributedApplicationModel>();
-
-    var resources = model.Resources
-        .Where(r => r.IsContainer() || r is ProjectResource || r is ExecutableResource)
-        .Where(r => r.Name != KnownResourceNames.AspireDashboard)
-        .ToList();
-
-    var commandTasks = new List<Task>();
-    foreach (var r in resources)
-    {
-        commandTasks.Add(commandService.ExecuteCommandAsync(r, commandName, cancellationToken));
-    }
-    await Task.WhenAll(commandTasks).ConfigureAwait(false);
-}
